Edit table dormitory

+++ database/migrations/2020_10_23_184305_create_table_dormitory.php
@@ -16,10 +16,10 @@ class CreateTableDormitory extends Migration
         Schema::create('dormitory', function (Blueprint $table) {
             $table->id();
             // $table->string('dormitory_name');
-            $table->integer('dormitory_unitsWater');
-            $table->integer('dormitory_serviceWater');
-            $table->integer('dormitory_unitselEctricity');
-            $table->integer('dormitory_serviceElectricity');
+            $table->double('dormitory_unitsWater');
+            $table->double('dormitory_serviceWater');
+            $table->double('dormitory_unitselEctricity');
+            $table->double('dormitory_serviceElectricity');
             $table->timestamps();
         });
     }
